Move the JWT DRF permissions into codecov_auth
This was created for webhooks, but we will want to have APIs that also can use this middleware.
So moving this to a more central place.
